- Removed -fno-strict-aliasing from the GCC flags for ZDoom and fixed the   issues that caused its inclusion. Is an optimized GCC build any faster   for being able to use strict aliasing rules? I dunno. It's still slower   than a VC++ build.
  I did run into two cases where TAutoSegIterator caused intractable problems
  with breaking strict aliasing rules, so I removed the templating from it,
  and the caller is now responsible for casting the probe value from void *.
- Removed #include "autosegs.h" from several files that did not need it
  (in particular, dobject.h when not compiling with VC++).


SVN r1743 (trunk)

+++ src/win32/i_dijoy.cpp
@@ -1293,7 +1293,7 @@ bool FDInputJoystickManager::IsXInputDeviceFast(const GUID *guid)
 			cbSize = rdi.cbSize = sizeof(rdi);
 			if (MyGetRawInputDeviceInfoA(devices[i].hDevice, RIDI_DEVICEINFO, &rdi, &cbSize) >= 0)
 			{
-				if(MAKELONG(rdi.hid.dwVendorId, rdi.hid.dwProductId) == guid->Data1)
+				if(MAKELONG(rdi.hid.dwVendorId, rdi.hid.dwProductId) == (LONG)guid->Data1)
 				{
 					char name[256];
 					UINT namelen = countof(name);
